+++ tests/tests_crashes_edgecases/test_noise_crashes.py
@@ -42,7 +42,7 @@ class TestNoiseCrashes:
             method_params=dict(abrupt_padding=28)
         )
 
-        # Expected segments to find a subset of based on current behavior
+        # Expected segments to find a subset of based on current behaviour
         expected_segments = [
             {'direction': 'Down', 'start': '2025-05-15', 'end': '2025-06-07'},
         ]
@@ -70,7 +70,7 @@ class TestNoiseCrashes:
             method_params=dict(abrupt_padding=28)
         )
 
-        # Expected segments to find a subset of based on current behavior
+        # Expected segments to find a subset of based on current behaviour
         expected_segments = [
             {'direction': 'Down', 'start': '2025-05-10', 'end': '2025-06-06'},
         ]
@@ -96,7 +96,7 @@ class TestNoiseCrashes:
             method_params=dict(abrupt_padding=28)
         )
         
-        # Expected segments to find a subset of based on current behavior
+        # Expected segments to find a subset of based on current behaviour
         expected_segments = [ 
             {'direction': 'Flat', 'start': '2025-03-31', 'end': '2025-04-10'}, 
             {'direction': 'Up', 'start': '2025-04-11', 'end': '2025-05-04'}, 
@@ -122,7 +122,7 @@ class TestNoiseCrashes:
             method_params=dict(abrupt_padding=28)
         )
         
-        # Expected segments to find a subset of based on current behavior
+        # Expected segments to find a subset of based on current behaviour
         expected_segments = [ 
             {'direction': 'Down', 'start': '2025-03-16', 'end': '2025-03-30'}, 
             {'direction': 'Up', 'start': '2025-04-03', 'end': '2025-05-04'}, 
@@ -149,7 +149,7 @@ class TestNoiseCrashes:
             method_params=dict(abrupt_padding=28)
         )
         
-        # Expected segments to find a subset of based on current behavior
+        # Expected segments to find a subset of based on current behaviour
         expected_segments = [ 
             {'direction': 'Up', 'start': '2025-04-12', 'end': '2025-05-06'}, 
             {'direction': 'Down', 'start': '2025-05-07', 'end': '2025-06-12'}, 
@@ -175,11 +175,11 @@ class TestNoiseCrashes:
             method_params=dict(abrupt_padding=28)
         )
         
-        # Expected segments to find a subset of based on current behavior
+        # Expected segments to find a subset of based on current behaviour
         expected_segments = [ 
-            {'direction': 'Up', 'start': '2025-04-11', 'end': '2025-04-29'}, 
-            {'direction': 'Flat', 'start': '2025-04-30', 'end': '2025-05-17'}, 
-            {'direction': 'Down', 'start': '2025-05-18', 'end': '2025-06-12'}, 
+            {'direction': 'Up', 'start': '2025-04-10', 'end': '2025-04-29'}, 
+            {'direction': 'Flat', 'start': '2025-04-30', 'end': '2025-05-16'}, 
+            {'direction': 'Down', 'start': '2025-05-17', 'end': '2025-06-12'}, 
         ]
         assert_segments_in_a_haystack(results.segments, expected_segments)
 
@@ -202,7 +202,7 @@ class TestNoiseCrashes:
             method_params=dict(abrupt_padding=28)
         )
         
-        # Expected segments to find a subset of based on current behavior
+        # Expected segments to find a subset of based on current behaviour
         expected_segments = [ 
             {'direction': 'Up', 'start': '2025-04-12', 'end': '2025-05-08'}, 
             {'direction': 'Down', 'start': '2025-05-09', 'end': '2025-06-17'}, 
@@ -229,9 +229,9 @@ class TestNoiseCrashes:
             method_params=dict(abrupt_padding=28)
         )
         
-        # Expected segments to find a subset of based on current behavior
+        # Expected segments to find a subset of based on current behaviour
         expected_segments = [ 
             {'direction': 'Up', 'start': '2025-03-31', 'end': '2025-05-09'}, #TODO: may expect some change if noise disabled
-            {'direction': 'Down', 'start': '2025-05-14', 'end': '2025-06-04'}, 
+            {'direction': 'Down', 'start': '2025-05-13', 'end': '2025-06-04'}, 
         ]
         assert_segments_in_a_haystack(results.segments, expected_segments)