+++ tests/tests_noise/test_noise_spikes_gradual.py
@@ -91,7 +91,7 @@ class TestNoiseSpikesGradual:
         
         # Expected noise segments representing the spike
         expected_noise_segments = [
-            {'direction': 'Noise', 'start': '2025-04-04', 'end': '2025-04-07'},
+            {'direction': 'Noise', 'start': '2025-04-05', 'end': '2025-04-07'},
         ]
         noise_segments = results.filter_segments(direction='Noise', format='dict')
         assert_segments_match(noise_segments, expected_noise_segments)
@@ -195,7 +195,7 @@ class TestNoiseSpikesGradual:
         
         # Expected noise segments representing the spike
         expected_noise_segments = [
-            {'direction': 'Noise', 'start': '2025-04-05', 'end': '2025-04-09'}, # TODO: later improve precision to be 04-07 start instead.
+            {'direction': 'Noise', 'start': '2025-04-07', 'end': '2025-04-09'},
         ]
         
         # Filter for noise segments and validate
@@ -452,6 +452,6 @@ class TestNoiseSpikesGradual:
             {'direction': 'Flat', 'start': '2025-05-05', 'end': '2025-05-06'},  
             {'direction': 'Flat', 'start': '2025-05-30', 'end': '2025-06-01'}, 
             {'direction': 'Flat', 'start': '2025-06-05', 'end': '2025-06-06'},
-            {'direction': 'Flat', 'start': '2025-06-18', 'end': '2025-06-30'},  # Last flat (not neighboring spike)
+            {'direction': 'Flat', 'start': '2025-06-18', 'end': '2025-06-30'},  # Last flat (not neighbouring spike)
         ]
         assert_segments_match(other_flats, expected_other_flats)